Transofrmations adjusted.

+++ PhongShadingCylinder/Transformations/CoordinateTranslator.cs
@@ -4,7 +4,10 @@ namespace PhongShadingCylinder.Transformations
 {
     public class CoordinateTranslator
     {
-        public static Vector2 Translate3dTo2d(Vector3 input, int width, int height)
+        /// <summary>
+        /// Translates the input vector from 3D scene coordinates to 2D (scene ceneter relative to top-left relative)
+        /// </summary>
+        public static Vector2 Translate3dTo2d(Vector3 input, float width, float height)
         {
             return new Vector2(width  / 2 * (1 + input.X),
                                height / 2 * (1 - input.Y));
